// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/filtroOmisosAusentismoBean.ts
export class FiltroOmisosAusentismoBean{
  idProceso: number;
  idEleccion: string;
  idAmbito: number;
  ambitoElectoral: string;
  codigoAmbitoElectoral: string;
  idCentroComputo: number;
  codigoCentroComputo: string;
  centroComputo: string;
  departamento: string;
  provincia: string;
  idUbigeo: number;
  mesa: string;
  proceso?: string;
  eleccion?: string;
  esquema?: string;
  distrito?:string;
  preferencial?: number;
}